requirements.txt
# Core dependencies
requests
requests-oauthlib
oauthlib
python-dotenv
pydantic
amadeus
mcp
uvicorn
